+++ src/user-lists/user-list-stats.controller.ts
@@ -7,6 +7,7 @@ import {
   ApiTags,
   ApiBadRequestResponse,
   ApiParam,
+  ApiQuery,
 } from "@nestjs/swagger";
 
 import { PageDto } from "../common/dtos/page.dto";
@@ -20,6 +21,8 @@ import { DbContributionStatTimeframe } from "./entities/contributions-timeframe.
 import { ContributionsTimeframeDto } from "./dtos/contributions-timeframe.dto";
 import { DbContributionsProjects } from "./entities/contributions-projects.entity";
 import { DbContributorCategoryTimeframe } from "./entities/contributors-timeframe.entity";
+import { ContributionsByProjectDto } from "./dtos/contributions-by-project.dto";
+import { TopProjectsDto } from "./dtos/top-projects.dto";
 
 @Controller("lists")
 @ApiTags("User Lists service")
@@ -74,8 +77,12 @@ export class UserListStatsController {
   @ApiNotFoundResponse({ description: "Unable to get contributions by project" })
   @ApiBadRequestResponse({ description: "Invalid request" })
   @ApiParam({ name: "id", type: "string" })
-  async getContributionsByProject(@Param("id") id: string): Promise<DbContributionsProjects[]> {
-    return this.userListStatsService.findContributionsByProject(id);
+  @ApiQuery({ name: "range", type: "integer", required: false })
+  async getContributionsByProject(
+    @Param("id") id: string,
+    @Query() options: ContributionsByProjectDto
+  ): Promise<DbContributionsProjects[]> {
+    return this.userListStatsService.findContributionsByProject(id, options);
   }
 
   @Get(":id/stats/top-project-contributions-by-contributor/")
@@ -91,9 +98,9 @@ export class UserListStatsController {
   @ApiParam({ name: "id", type: "string" })
   async getContributorContributionsByProject(
     @Param("id") id: string,
-    @Query("repoId") repoId: number
+    @Query() options: TopProjectsDto
   ): Promise<DbUserListContributorStat[]> {
-    return this.userListStatsService.findListContributorStatsByProject(id, repoId);
+    return this.userListStatsService.findListContributorStatsByProject(options, id);
   }
 
   @Get(":id/stats/contributions-evolution-by-contributor-type")